customizable_navbar/admin/containers: refactor folder structure
