+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/HiddenNavbarItemList/HiddenNavbarItemList.test.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { render, screen, fireEvent } from 'utils/testUtils/rtl';
 import HiddenNavbarItemList from '.';
 import allNavbarItems from 'hooks/fixtures/navbarItems';
-import { addNavbarItem } from '../../../services/navbar';
-import { deletePage } from '../../../services/pages';
-import { NAVIGATION_PATH } from '../';
+import { addNavbarItem } from '../../../../services/navbar';
+import { deletePage } from '../../../../services/pages';
+import { NAVIGATION_PATH } from '../..';
 import clHistory from 'utils/cl-router/history';
 
 jest.mock('services/locale');
@@ -14,7 +14,7 @@ let mockNavbarItems = allNavbarItems;
 const mockRemovedDefaultNavbarItems = [];
 
 jest.mock('hooks/useNavbarItems', () => jest.fn(() => mockNavbarItems));
-jest.mock('../../../hooks/useRemovedDefaultNavbarItems', () =>
+jest.mock('../../../../hooks/useRemovedDefaultNavbarItems', () =>
   jest.fn(() => mockRemovedDefaultNavbarItems)
 );
 
@@ -22,12 +22,12 @@ jest.mock('hooks/usePages');
 jest.mock('hooks/usePageSlugById');
 jest.mock('hooks/useLocale');
 
-jest.mock('../../../services/navbar', () => ({
+jest.mock('../../../../services/navbar', () => ({
   addNavbarItem: jest.fn(),
 }));
 
-jest.mock('../../../services/pages', () => {
-  const original = jest.requireActual('../../../services/pages');
+jest.mock('../../../../services/pages', () => {
+  const original = jest.requireActual('../../../../services/pages');
 
   return {
     ...original,
