+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/HiddenNavbarItemList/index.tsx
@@ -1,33 +1,33 @@
 import React, { useMemo } from 'react';
 
 // services
-import { addNavbarItem } from '../../../services/navbar';
-import { deletePage } from '../../../services/pages';
+import { addNavbarItem } from '../../../../services/navbar';
+import { deletePage } from '../../../../services/pages';
 import { IPageData, FIXED_PAGES, TPageCode } from 'services/pages';
 import { getNavbarItemSlug } from 'services/navbar';
 
 // hooks
 import useNavbarItems from 'hooks/useNavbarItems';
-import useRemovedDefaultNavbarItems from '../../../hooks/useRemovedDefaultNavbarItems';
+import useRemovedDefaultNavbarItems from '../../../../hooks/useRemovedDefaultNavbarItems';
 import usePages from 'hooks/usePages';
 import usePageSlugById from 'hooks/usePageSlugById';
 import useLocale from 'hooks/useLocale';
 
 // components
 import { List, Row } from 'components/admin/ResourceList';
-import NavbarItemRow from '../../components/NavbarItemRow';
+import NavbarItemRow from '../../../components/NavbarItemRow';
 import Header from './Header';
 
 // i18n
 import { injectIntl } from 'utils/cl-intl';
 import { InjectedIntlProps } from 'react-intl';
-import messages from '../messages';
+import messages from '../../messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
 import getItemsNotInNavbar, { IItemNotInNavbar } from './getItemsNotInNavbar';
 import clHistory from 'utils/cl-router/history';
-import { NAVIGATION_PATH } from '../';
+import { NAVIGATION_PATH } from '../..';
 
 const FIXED_PAGES_SET = new Set<TPageCode>(FIXED_PAGES);
 const removeFixedPages = (page: IPageData) =>
